Add files via upload
#CENSUS
Q. 1) How will you hide the indexes of the dataframe.
Q. 2) How can we set the caption / heading on the dataframe.
Q. 3) Show the records related with the districts - New Delhi , Lucknow , Jaipur.
Q. 4) Calculate state-wise : A. Total number of population. B. Total no. of the population with different religions. 
Q. 5) How many Male Workers were there in Maharashtra state ? 
Q. 6) How to set a column as index of the dataframe ?
Q. 7a) Add a Suffix to the column names. Q. 7b) Add a Prefix to the column names.